add error handling

+++ Application/Activities/Delete.cs
@@ -1,3 +1,4 @@
+using Application.Core;
 using AutoMapper;
 using Domain;
 using MediatR;
@@ -7,12 +8,12 @@ namespace Application.Activities;
 
 public class Delete
 {
-    public class Command : IRequest
+    public class Command : IRequest<Result<Unit>>
     {
         public Guid Id { get; set; }
     }
     
-    public class Handler : IRequestHandler<Command>
+    public class Handler : IRequestHandler<Command, Result<Unit>>
     {
         private readonly DataContext _context;
 
@@ -21,13 +22,19 @@ public class Delete
             _context = context;
         }
 
-        public async Task Handle(Command request, CancellationToken cancellationToken)
+        public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
         {
             var activity = await _context.Activities.FindAsync(request.Id);
 
+            if (activity == null) return null;
+
             _context.Activities.Remove(activity);
-            
-            await _context.SaveChangesAsync();
+
+            var result = await _context.SaveChangesAsync() > 0;
+            if (!result)
+                return Result<Unit>.Failure("Failed to delete the activity");
+
+            return Result<Unit>.Success(Unit.Value);
         }
     }
